Create HuanLeSongQuiz
Program contains one known error (needs to be revised):

Incorrectly displays "Error. Not a valid input, please enter valid input" if the wrong answer is put in, rather than displaying the message if a character other than A, B or C is put.

Fix issue so user is required to answer the question again if an invalid character is inputted. The message should also not be displayed if the user selects the wrong answer.

2025-06-08